refactor(geotiff): extract GeoTIFF fixture builder
Consolidate duplicated hand-built TIFF helper functions into a single
parameterised `write_minimal_tiff` utility in `_geotiff_fixtures.py`.

This replaces the redundant implementations in `test_geotags.py` and
`test_metadata.py`, reducing code duplication and improving
maintainability of the GeoTIFF test suite.
